# app/api/download/route.ts
import { NextResponse, type NextRequest } from "next/server";
import { createPublicRoute } from "@/lib/api-middleware";
import { getAccessToken, getFileDetailsFromDrive } from "@/lib/drive";
import { logActivity } from "@/lib/activityLogger";
import { trackBandwidth } from "@/lib/analyticsTracker";
import { logger } from "@/lib/logger";
import { ERROR_MESSAGES } from "@/lib/constants";
import { kv } from "@/lib/kv";
import { db } from "@/lib/db";
import {
  validateDownloadRequest,
  prepareGoogleDriveUrl,
  prepareResponseHeaders,
} from "@/lib/services/download";
import { applyWatermark } from "@/lib/watermark";

export const dynamic = "force-dynamic";

function applyPreviewResponseHeaders(headers: Headers) {
  const disposition = headers.get("Content-Disposition");
  headers.set(
    "Content-Disposition",
    disposition ? disposition.replace(/^attachment\b/i, "inline") : "inline",
  );
  headers.set("Cache-Control", "no-store, no-cache, must-revalidate");
  headers.set("Pragma", "no-cache");
  headers.set("Expires", "0");
}

async function handleDownload(request: NextRequest) {
  try {
    const { context, session, error } = await validateDownloadRequest(request);
    if (error) {
      return NextResponse.json(
        { error: error.error },
        { status: error.status },
      );
    }

    const { fileId, range } = context;
    const exportFormat =
      request.nextUrl.searchParams.get("export") === "pdf" ? "pdf" : null;
    const isPreview = request.nextUrl.searchParams.get("preview") === "1";
    logger.info({ fileId }, "[Download] Starting download");

    if (fileId.startsWith("local-storage:")) {
      const { getDownloadStream } = await import("@/lib/storage");
      try {
        const downloadData = await getDownloadStream(fileId);

        if (!downloadData) {
          return NextResponse.json(
            { error: ERROR_MESSAGES.FILE_NOT_FOUND },
            { status: 404 },
          );
        }

        const { stream: webStream, size, mimeType, filename } = downloadData;

        let finalBody: BodyInit = webStream as ReadableStream<Uint8Array>;
        let finalSize = size;

        if (
          context.shareRecord?.hasWatermark &&
          !range &&
          (mimeType.startsWith("image/") || mimeType === "application/pdf")
        ) {
          const response = new Response(webStream);
          const arrayBuffer = await response.arrayBuffer();
          const buffer = Buffer.from(arrayBuffer);
          const watermarked = await applyWatermark(
            buffer,
            mimeType,
            context.shareRecord.watermarkText || "",
          );

          if (watermarked) {
            finalBody = new Blob([watermarked as unknown as BlobPart]);
            finalSize = watermarked.length;
          }
        }

        const responseHeaders = prepareResponseHeaders(
          mimeType,
          filename,
          range,
          request.headers.get("Sec-Fetch-Dest"),
          null,
          false,
          request.headers.get("origin"),
        );
        if (isPreview) {
          applyPreviewResponseHeaders(responseHeaders);
        }
        responseHeaders.set("Content-Length", finalSize.toString());
        responseHeaders.set("Accept-Ranges", "bytes");

        return new Response(finalBody, {
          status: 200,
          headers: responseHeaders,
        });
      } catch (err: unknown) {
        if (
          err instanceof Error &&
          err.message === "Cannot download a directory"
        ) {
          return NextResponse.json(
            { error: ERROR_MESSAGES.FOLDER_DOWNLOAD_NOT_SUPPORTED },
            { status: 400 },
          );
        }
        throw err;
      }
    }

    const [accessToken, fileDetails] = await Promise.all([
      getAccessToken(),
      getFileDetailsFromDrive(fileId),
    ]);

    if (!fileDetails) {
      return NextResponse.json(
        { error: ERROR_MESSAGES.FILE_NOT_FOUND },
        { status: 404 },
      );
    }

    if (fileDetails.mimeType === "application/vnd.google-apps.folder") {
      return NextResponse.json(
        { error: ERROR_MESSAGES.FOLDER_DOWNLOAD_NOT_SUPPORTED },
        { status: 400 },
      );
    }

    const { url, mimeType, filename } = prepareGoogleDriveUrl(
      fileId,
      fileDetails,
      exportFormat,
    );

    const googleRequestHeaders = new Headers();
    googleRequestHeaders.set("Authorization", `Bearer ${accessToken}`);
    googleRequestHeaders.set("User-Agent", "NOT-CLOUD-Streamer/1.0");
    googleRequestHeaders.set("Accept-Encoding", "identity");

    const isGoogleDoc = fileDetails.mimeType.startsWith(
      "application/vnd.google-apps.",
    );

    if (range && !isGoogleDoc) {
      googleRequestHeaders.set("Range", range);
    }

    if (request.method === "HEAD") {
      const headHeaders = prepareResponseHeaders(
        mimeType,
        filename,
        null,
        null,
        null,
        true,
        request.headers.get("origin"),
      );
      if (fileDetails.size) headHeaders.set("Content-Length", fileDetails.size);
      headHeaders.set("Accept-Ranges", "bytes");
      if (isPreview) {
        applyPreviewResponseHeaders(headHeaders);
      }
      return new Response(new Blob([]), { status: 200, headers: headHeaders });
    }

    const controller = new AbortController();
    const downloadTimeout = parseInt(
      process.env.DOWNLOAD_TIMEOUT_MS || "60000",
      10,
    );
    const timeoutId = setTimeout(() => controller.abort(), downloadTimeout);

    const googleResponse = await fetch(url, {
      headers: googleRequestHeaders,
      method: "GET",
      cache: "no-store",
      signal: controller.signal,
    }).finally(() => clearTimeout(timeoutId));

    if (!googleResponse.ok || googleResponse.status === 204) {
      const errorJson =
        googleResponse.status !== 204
          ? await googleResponse.json().catch(() => ({}))
          : {};
      logger.error(
        { status: googleResponse.status, errorJson, fileId },
        "Google Drive API Error",
      );
      return NextResponse.json(
        {
          error:
            errorJson.error?.message ||
            (googleResponse.status === 204
              ? "Google Drive mengembalikan respons kosong. Mungkin token akses kedaluwarsa."
              : "Gagal mengambil file dari Google Drive"),
        },
        { status: googleResponse.status === 204 ? 502 : googleResponse.status },
      );
    }

    const responseHeaders = prepareResponseHeaders(
      mimeType,
      filename,
      range,
      request.headers.get("Sec-Fetch-Dest"),
      googleResponse,
      false,
      request.headers.get("origin"),
    );
    if (isPreview) {
      applyPreviewResponseHeaders(responseHeaders);
    }

    if (!range) {
      const forwardedFor = request.headers.get("x-forwarded-for");
      const realIp = request.headers.get("x-real-ip");
      const ip = forwardedFor
        ? forwardedFor.split(",")[0].trim()
        : realIp || "anonymous";
      const userIdentifier = session?.user?.email || ip;

      const dedupeKey = `loop_prevent:download:${fileId}:${userIdentifier}`;
      const isDuplicate = await kv.get(dedupeKey);

      if (!isDuplicate) {
        await kv.set(dedupeKey, "1", { ex: 5 });

        if (context.shareRecord) {
          try {
            await db.shareLink.update({
              where: { jti: context.shareRecord.jti },
              data: { views: { increment: 1 } },
            });
          } catch (e) {
            logger.error({ err: e }, "Failed to increment sharelink uses");
          }
        }

        logActivity("DOWNLOAD", {
          itemName: fileDetails.name,
          itemId: fileId,
          itemSize: fileDetails.size || "0",
          userEmail: session?.user?.email,
          metadata: {
            fileId,
            mimeType,
            rangeRequest: false,
            isShareAccess: !!context.shareRecord,
            shareLinkId: context.shareRecord?.jti,
          },
        }).catch((e) =>
          logger.error({ err: e }, "Gagal mencatat log aktivitas"),
        );

        const downloadSize = parseInt(fileDetails.size || "0", 10);
        if (downloadSize > 0) {
          trackBandwidth(downloadSize).catch((err) =>
            logger.warn({ err, fileId }, "Failed to track bandwidth"),
          );
        }
      } else {
        logger.info(
          { fileId, userIdentifier },
          "[Download] Skipping duplicate log",
        );
      }
    }

    let finalBody: BodyInit | null = googleResponse.body;

    if (
      context.shareRecord?.hasWatermark &&
      !range &&
      (mimeType.startsWith("image/") || mimeType === "application/pdf")
    ) {
      const arrayBuffer = await googleResponse.arrayBuffer();
      const buffer = Buffer.from(arrayBuffer);
      const watermarked = await applyWatermark(
        buffer,
        mimeType,
        context.shareRecord.watermarkText || "",
      );

      if (watermarked) {
        finalBody = new Blob([watermarked as unknown as BlobPart]);
        responseHeaders.set("Content-Length", watermarked.length.toString());
      } else {
        finalBody = new Blob([buffer as unknown as BlobPart]);
      }
    }

    return new Response(finalBody, {
      status: googleResponse.status,
      headers: responseHeaders,
    });
  } catch (error: unknown) {
    logger.error({ err: error }, "Download API Error");
    return NextResponse.json(
      {
        error:
          error instanceof Error
            ? error.message
            : ERROR_MESSAGES.INTERNAL_SERVER_ERROR,
      },
      { status: 500 },
    );
  }
}

export const GET = createPublicRoute(
  async ({ request }) => {
    return await handleDownload(request);
  },
  { rateLimit: false },
);

export const HEAD = createPublicRoute(
  async ({ request }) => {
    return await handleDownload(request);
  },
  { rateLimit: false },
);