# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NOT CLOUD API
  description: |
    Core API documentation for NOT CLOUD - a self-hosted Google Drive CMS,
    explorer, and streaming platform.

    ## Authentication
    Most endpoints support one of:
    - **sessionAuth** — NextAuth.js JWT session cookie (`next-auth.session-token`)
    - **apiKeyAuth** — `Authorization: Bearer <zk_xxx>` header (see API Keys section)
    - **shareToken** — `share_token` query parameter for shared links
    - **accessToken** — `access_token` query parameter for protected folders

    ## Rate Limiting
    - General API: 500 requests per minute
    - Download API: 100 requests per hour
    - Auth API: 20 requests per 5 minutes
    - Admin API: 50 requests per minute
    - **API Key**: 1000 requests per minute

    ## Error Responses
    Most errors follow this format:
    ```json
    {
      "error": "Error message"
    }
    ```
  version: 2.0.0
  contact:
    name: NOT CLOUD Support
    url: https://www.notnot.store/
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://zee-index.duckdns.org
    description: Production server
...
paths:
...
  /api/download:
    get:
      tags:
        - Files
      summary: Download a file
      description: Initiate file download from Google Drive
      operationId: downloadFile
      parameters:
        - name: fileId
          in: query
          required: true
          schema:
            type: string
        - name: share_token
          in: query
          schema:
            type: string
        - name: access_token
          in: query
          schema:
            type: string
          description: Folder access token for protected folders
      responses:
        "200":
          description: File stream
        "302":
          description: Redirect to Google Drive download URL
        "401":
          description: Unauthorized
        "404":
          description: File not found
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: zk_
      description: API key with bearer format (zk_ prefix). Create one from the admin panel.
...